+++ app/api/urls.py
@@ -1,7 +1,6 @@
-from django.conf.urls import url
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import SchoolViewSet, EntryViewSet, SpecialtyViewSet, SchoolEntryList
+from api.views import SchoolViewSet, EntryViewSet, SpecialtyViewSet, SchoolEntryList
 
 # Create a router and register our viewsets with it.
 router = DefaultRouter()
