+++ app/api/views.py
@@ -4,7 +4,7 @@ from rest_framework import permissions
 from schools.models import School
 from users.models import Profile
 from main_app.models import Entry, Specialty
-from .serializers import SchoolSerializer, EntrySerializer, SpecialtySerializer
+from api.serializers import SchoolSerializer, EntrySerializer, SpecialtySerializer
 
 
 class SpecialtyViewSet(viewsets.ReadOnlyModelViewSet):
