+++ app/main_app/urls.py
@@ -29,5 +29,4 @@ urlpatterns = [
     path('about/', views.about, name='about'),
     path('help/', views.help, name='help'),
     path('specialties/', SpecialtiesListView.as_view(), name='specialties'),
-    path('api/', include('main_app.api.urls')),
 ]
