+++ app/vacancies/urls.py
@@ -21,6 +21,7 @@ from django.contrib.auth import views as auth_views
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
+    path('api/', include('api.urls')),
     path('maskeparty/', include('impersonate.urls')),
     path('', include('main_app.urls')),
     path('', include('history.urls')),
